+++ android/app/src/main/java/com/xianxia/sect/core/engine/GameEngineCore.kt
@@ -46,6 +46,7 @@ class GameEngineCore @Inject constructor(
     private val stateStore: GameStateStore,
     private val stateManager: UnifiedGameStateManager,
     private val eventBus: EventBus,
+    @Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
     private val performanceMonitor: GamePerformanceMonitor,
     private val systemManager: SystemManager
 ) {