+++ android/app/src/main/java/com/xianxia/sect/core/util/GameMonitorManager.kt
@@ -24,6 +24,7 @@ private class GameMemoryPressureListener : XianxiaApplication.MemoryPressureList
     override fun onLowMemory() { /* handled by manager */ }
 }
 
+@Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
 @Singleton
 class GameMonitorManager @Inject constructor(
     private val memoryMonitor: MemoryMonitor,
@@ -181,6 +182,7 @@ class GameMonitorManager @Inject constructor(
     
     fun getMemoryMonitor(): MemoryMonitor = memoryMonitor
     
+    @Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
     fun getPerformanceMonitor(): PerformanceMonitor = performanceMonitor
     
     fun getGCOptimizer(): GCOptimizer = gcOptimizer
@@ -219,6 +221,7 @@ class GameMonitorManager @Inject constructor(
         unifiedPerformanceMonitor.removeListener(listener)
     }
     
+    @Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
     fun isPerformanceHealthy(): Boolean {
         val memoryInfo = memoryMonitor.getCurrentMemoryInfo()
         val isPerformanceOk = performanceMonitor.isPerformanceAcceptable()
@@ -226,6 +229,7 @@ class GameMonitorManager @Inject constructor(
         return isPerformanceOk && (memoryInfo == null || !memoryInfo.isCritical)
     }
     
+    @Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
     fun getOptimizationRecommendation(): OptimizationRecommendation {
         val memoryInfo = memoryMonitor.getCurrentMemoryInfo()
         val perfLevel = performanceMonitor.getRecommendedOptimizationLevel()
@@ -240,6 +244,7 @@ class GameMonitorManager @Inject constructor(
         )
     }
     
+    @Suppress("DEPRECATION") // TODO: Migrate to UnifiedPerformanceMonitor (P1/P3 task)
     data class OptimizationRecommendation(
         val performanceLevel: PerformanceMonitor.OptimizationLevel,
         val recommendedGCType: GCOptimizer.GCType,